+++ relay-dynamic-config/src/global.rs
@@ -1,7 +1,9 @@
 use std::collections::HashMap;
 use std::fs::File;
 use std::io::BufReader;
+use std::iter::Chain;
 use std::path::Path;
+use std::slice::Iter;
 
 use relay_event_normalization::MeasurementsConfig;
 use relay_quotas::Quota;
@@ -23,7 +25,10 @@ pub struct GlobalConfig {
     #[serde(skip_serializing_if = "Vec::is_empty")]
     pub quotas: Vec<Quota>,
     /// Sentry options passed down to Relay.
-    #[serde(skip_serializing_if = "is_default")]
+    #[serde(
+        deserialize_with = "default_on_error",
+        skip_serializing_if = "is_default"
+    )]
     pub options: Options,
 }
 
@@ -53,6 +58,7 @@ pub struct Options {
     /// of improving profile (function) metrics
     #[serde(
         rename = "profiling.profile_metrics.unsampled_profiles.platforms",
+        deserialize_with = "default_on_error",
         skip_serializing_if = "Vec::is_empty"
     )]
     pub profile_metrics_allowed_platforms: Vec<String>,
@@ -60,6 +66,7 @@ pub struct Options {
     /// Sample rate for tuning the amount of unsampled profiles that we "let through"
     #[serde(
         rename = "profiling.profile_metrics.unsampled_profiles.sample_rate",
+        deserialize_with = "default_on_error",
         skip_serializing_if = "is_default"
     )]
     pub profile_metrics_sample_rate: f32,
@@ -67,6 +74,7 @@ pub struct Options {
     /// Kill switch for shutting down profile metrics
     #[serde(
         rename = "profiling.profile_metrics.unsampled_profiles.enabled",
+        deserialize_with = "default_on_error",
         skip_serializing_if = "is_default"
     )]
     pub unsampled_profiles_enabled: bool,
@@ -74,6 +82,7 @@ pub struct Options {
     /// Kill switch for controlling the cardinality limiter.
     #[serde(
         rename = "relay.cardinality-limiter.mode",
+        deserialize_with = "default_on_error",
         skip_serializing_if = "is_default"
     )]
     pub cardinality_limiter_mode: CardinalityLimiterMode,
@@ -84,6 +93,7 @@ pub struct Options {
     /// If set to `1.0` all cardinality limiter rejections will be logged as a Sentry error.
     #[serde(
         rename = "relay.cardinality-limiter.error-sample-rate",
+        deserialize_with = "default_on_error",
         skip_serializing_if = "is_default"
     )]
     pub cardinality_limiter_error_sample_rate: f32,
@@ -91,7 +101,11 @@ pub struct Options {
     /// Kill switch for disabling the span usage metric.
     ///
     /// This metric is converted into outcomes in a sentry-side consumer.
-    #[serde(rename = "relay.span-usage-metric", skip_serializing_if = "is_default")]
+    #[serde(
+        rename = "relay.span-usage-metric",
+        deserialize_with = "default_on_error",
+        skip_serializing_if = "is_default"
+    )]
     pub span_usage_metric: bool,
 
     /// All other unknown options.
@@ -148,9 +162,6 @@ impl<'a> DynamicQuotas<'a> {
     }
 }
 
-use std::iter::Chain;
-use std::slice::Iter;
-
 // Implementing IntoIterator for &DynamicQuotas to allow iterating over the quotas.
 impl<'a> IntoIterator for DynamicQuotas<'a> {
     type Item = &'a Quota;
@@ -158,6 +169,24 @@ impl<'a> IntoIterator for DynamicQuotas<'a> {
 
     fn into_iter(self) -> Self::IntoIter {
         self.global_quotas.iter().chain(self.project_quotas.iter())
+    }
+}
+
+fn default_on_error<'de, D, T>(deserializer: D) -> Result<T, D::Error>
+where
+    D: serde::de::Deserializer<'de>,
+    T: Default + serde::de::DeserializeOwned,
+{
+    match T::deserialize(deserializer) {
+        Ok(value) => Ok(value),
+        Err(error) => {
+            relay_log::error!(
+                error = %error,
+                "Error deserializing global config option: {}",
+                std::any::type_name::<T>(),
+            );
+            Ok(T::default())
+        }
     }
 }
 
@@ -224,6 +253,21 @@ mod tests {
         for (expected_id, quota) in ["foo", "bar", "baz", "qux"].iter().zip(dynamic_quotas) {
             assert_eq!(Some(expected_id.to_string()), quota.id);
         }
+    }
+
+    #[test]
+    fn test_global_config_invalid_value_is_default() {
+        let options: Options = serde_json::from_str(
+            r#"{"relay.cardinality-limiter.mode":"passive","relay.span-usage-metric":123}"#,
+        )
+        .unwrap();
+
+        let expected = Options {
+            cardinality_limiter_mode: CardinalityLimiterMode::Passive,
+            ..Default::default()
+        };
+
+        assert_eq!(options, expected);
     }
 
     #[test]